+++ index.js
@@ -1,8 +1,8 @@
-export { Mockaton } from './src/Mockaton.js'
-export { Commander } from './src/ApiCommander.js'
+export { Mockaton } from './src/server/Mockaton.js'
+export { Commander } from './src/client/ApiCommander.js'
 
-export { jwtCookie } from './src/utils/jwt.js'
-export { jsToJsonPlugin } from './src/MockDispatcher.js'
-export { parseJSON, BodyReaderError } from './src/utils/http-request.js'
+export { jwtCookie } from './src/server/utils/jwt.js'
+export { jsToJsonPlugin } from './src/server/MockDispatcher.js'
+export { parseJSON, BodyReaderError } from './src/server/utils/http-request.js'
 
 export const defineConfig = opts => opts