+++ src/client/app.js
@@ -1,7 +1,7 @@
-import { createElement as r, createSvgElement as s, className, restoreFocus, Defer, Fragment } from './DashboardDom.js'
+import { createElement as r, createSvgElement as s, className, restoreFocus, Defer, Fragment } from './dom-utils.js'
 import { HEADER_502 } from './ApiConstants.js'
 import { parseFilename } from './Filename.js'
-import { store } from './DashboardStore.js'
+import { store } from './app-store.js'
 
 
 const CSS = {
@@ -97,7 +97,7 @@ function Header() {
 	return (
 		r('header', null,
 			r('object', {
-				data: 'Logo.svg',
+				data: 'logo.svg',
 				type: 'image/svg+xml',
 				width: 120,
 				height: 22