+++ src/server/Mockaton.test.js
@@ -9,15 +9,17 @@ import { describe, test, before, beforeEach, after } from 'node:test'
 import { mkdtempSync } from 'node:fs'
 import { writeFile, unlink, mkdir, readFile, rename } from 'node:fs/promises'
 
-import { API } from './ApiConstants.js'
 import { logger } from './utils/logger.js'
 import { mimeFor } from './utils/mime.js'
-import { Mockaton } from './Mockaton.js'
 import { readBody } from './utils/http-request.js'
-import { Commander } from './ApiCommander.js'
 import { CorsHeader } from './utils/http-cors.js'
+
+import { API } from './ApiConstants.js'
+import { Mockaton } from './Mockaton.js'
 import { parseFilename } from './Filename.js'
 import { watchMocksDir, watchStaticDir } from './Watcher.js'
+
+import { Commander } from '../client/ApiCommander.js'
 
 
 const mocksDir = mkdtempSync(join(tmpdir(), 'mocks'))